Storing templates in-memory to avoid re-generating
the results of template.ParseFiles is stored in a variable templates.
This improves the performance of the code by eliminating the need to
read view.html and edit.html from disk on each request.

The cost of this change is that the server must be restarted in order to
reload changes to view.html and edit.html, where previously they would
be reloaded on next request.